feat: restructure javascript_common package
* to allow lower-level sharing of common/lib
