+++ packages/javascript_common/src/lib/index.ts
@@ -1,5 +1,7 @@
+export * from './interfaces'
 export * from './asyncify.js';
 export * from './async_mutex.js';
 export * from './handle_map.js';
 export * from './stream.js';
 export * from './wasm.js';
+export * as sf_host from './sf_host.js';
