+++ packages/javascript_common/src/lib/sf_host.ts
@@ -1,5 +1,6 @@
-import type { TextCoder, AppContext } from './interfaces';
-import { Asyncify, AsyncifyState, HandleMap } from './lib/index.js';
+import type { TextCoder, AppContext, AppContextSync } from '../interfaces';
+import { Asyncify, AsyncifyState } from './asyncify.js'
+import { HandleMap } from './handle_map.js'
 import { WasiErrno } from './error.js';
 
 type AbiResult = number;
@@ -104,9 +105,7 @@ export function link(app: AppContext, textCoder: TextCoder, asyncify: Asyncify):
     return app.writeStream(handle, data).then(c => abi_ok(c), e => abi_err(e.errno));
   }
   async function __export_stream_close(handle: number): Promise<AbiResult> {
-    await app.closeStream(handle).then(_ => abi_ok(0), e => abi_err(e.errno));
-
-    return abi_ok(0);
+    return await app.closeStream(handle).then(_ => abi_ok(0), e => abi_err(e.errno))
   }
 
   return {
@@ -116,6 +115,82 @@ export function link(app: AppContext, textCoder: TextCoder, asyncify: Asyncify):
       stream_read: asyncify.wrapImport(__export_stream_read, 0),
       stream_write: asyncify.wrapImport(__export_stream_write, 0),
       stream_close: asyncify.wrapImport(__export_stream_close, 0)
+    }
+  }
+}
+
+export function linkSync(app: AppContextSync, textCoder: TextCoder): WebAssembly.Imports {
+  const messageStore = new HandleMap<ArrayBuffer>();
+
+  function __export_message_exchange(msg_ptr: Ptr<8>, msg_len: Size, out_ptr: Ptr<8>, out_len: Size, ret_handle: Ptr<32>): Size {
+    const msg = JSON.parse(textCoder.decodeUtf8(
+      app.memoryBytes.subarray(msg_ptr, msg_ptr + msg_len)
+    ));
+    const response = app.handleMessage(msg);
+
+    let messageHandle = 0;
+    const responseBytes = textCoder.encodeUtf8(JSON.stringify(response));
+    if (responseBytes.byteLength > out_len) {
+      messageHandle = messageStore.insert(responseBytes);
+    } else {
+      const out = app.memoryBytes.subarray(out_ptr, out_ptr + out_len);
+      writeBytes(responseBytes, out);
+    }
+
+    app.memoryView.setInt32(ret_handle, messageHandle, true);
+    return responseBytes.byteLength;
+  }
+  function __export_message_exchange_retrieve(handle: number, out_ptr: Ptr<8>, out_len: Size): AbiResult {
+    const responseBytes = messageStore.remove(handle);
+    if (responseBytes === undefined) {
+      return abi_err(WasiErrno.EBADF);
+    }
+    if (out_len < responseBytes.byteLength) {
+      return abi_err(WasiErrno.ERANGE);
+    }
+
+    const out = app.memoryBytes.subarray(out_ptr, out_ptr + out_len);
+    writeBytes(responseBytes, out);
+
+    return abi_ok(responseBytes.byteLength);
+  }
+
+  function __export_stream_read(handle: number, out_ptr: Ptr<8>, out_len: Size): AbiResult {
+    const out = app.memoryBytes.subarray(out_ptr, out_ptr + out_len);
+
+    try {
+      const c = app.readStream(handle, out)
+      return abi_ok(c)
+    } catch (e: any) {
+      return abi_err(e.errno)
+    }
+  }
+  function __export_stream_write(handle: number, in_ptr: Ptr<8>, in_len: Size): AbiResult {
+    const data = app.memoryBytes.subarray(in_ptr, in_ptr + in_len);
+
+    try {
+      const c = app.writeStream(handle, data)
+      return abi_ok(c)
+    } catch (e: any) {
+      return abi_err(e.errno)
+    }
+  }
+  function __export_stream_close(handle: number): AbiResult {
+    try {
+      app.closeStream(handle)
+      return abi_ok(0)
+    } catch (e: any) {
+      return abi_err(e.errno)
+    }
+  }
+
+  return {
+    sf_host_unstable: {
+      message_exchange: __export_message_exchange,
+      message_exchange_retrieve: __export_message_exchange_retrieve,
+      stream_read: __export_stream_read,
+      stream_write: __export_stream_write,
+      stream_close: __export_stream_close,
     }
   }
 }
